% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq-plots.R
\name{revlog_trans}
\alias{revlog_trans}
\title{Reverse logarithm transformation}
\usage{
revlog_trans(base = exp(1))
}
\arguments{
\item{base}{The base with respect to which to take the logarithm (defaults to e).}
}
\value{
A scale transformation object.
}
\description{
Reverse logarithm transformation
}
