% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq-plots.R
\name{stat_qq_band}
\alias{stat_qq_band}
\title{Add confidence band for a QQ plot}
\usage{
stat_qq_band(
  mapping = NULL,
  data = NULL,
  geom = "ribbon",
  position = "identity",
  show.legend = FALSE,
  inherit.aes = TRUE,
  distribution = "unif",
  max_pts_to_plot = 500,
  ci_level = 0.95,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{distribution}{The distribution with respect to which to make QQ plot
(e.g. \code{"unif"}, \code{"norm"}, etc.); default \code{"unif"}}

\item{max_pts_to_plot}{Points are downsampled for plotting purposes, so that
the maximum number of points actually plotted is \code{max_pts_to_plot}.}

\item{ci_level}{The pointwise confidence level for the QQ plot; default 0.95}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Add confidence band for a QQ plot
}
